API UPDATE

+++ SistemaJuridicoWebAPI/SistemaJuridicoWebAPI/Models/PROCESSO_ACORDO.cs
@@ -6,8 +6,8 @@ namespace SistemaJuridicoWebAPI.Models
   {
     [Key]
     public Guid ID { get; set; }
-    public string DATA_ACORDO { get; set; }
-    public int VALOR_ACORDO { get; set; }
+    public string  DATA_ACORDO { get; set; }
+    public float VALOR_ACORDO { get; set; }
     public string ID_PROCESSO { get; set; }
     public string CRIADOR_ACORDO { get; set; }
     public string CONDICOES_TENTATIVA_DE_ACORDO { get; set; }
